requirements.txt
snowflake-connector-python[pandas]
python-dotenv
pandas
streamlit
pyyaml